media/requirements.txt
fastapi
minio
python-dotenv